Split out process system calls in the libc backend.
